+++ eosets/__meta__.py
@@ -17,4 +17,4 @@
 **EOSets** library
 """
 
-__version__ = "0.3.3"
+__version__ = "0.3.4.dev0"